backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
google-generativeai
psycopg2-binary
pydantic
pydantic[email]
bcrypt
python-multipart
alembic